requirements.txt
Flask==2.0.3
werkzeug==2.0.3
gunicorn==20.1.0
ffmpeg-python==0.2.0
Flask-WTF==1.0.0
Flask-Limiter==2.0.2
Flask-Talisman==1.0.0
cryptography==39.0.1